Use locking Queue
Sometimes the sounds are played out of order.  This is probably caused
by the end of stream message handler being called in the GStreamer
thread; a Python list is not interlocked.

Change to using the Python synchronised queue class.  Care is taken not
to call Queue.get() if the queue is empty, as it would block.